requirements.txt
# Core ML
xgboost>=1.7.0
scikit-learn>=1.2.0
numpy>=1.23.0
pandas>=1.5.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0

# Logging
loguru>=0.6.0

# Reproducibility
python-dotenv>=0.21.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0